Extract timer code into its own timer/ package
- Move Timer interface, ProcessLocalTimer, WebTimer, GetTimers,
  RunTimer, RunBreakTimer into github.com/remotemobprogramming/mob/v5/timer
- WebTimer now holds Room and TimerUser as struct fields (resolved by
  main before constructing the timer)
- ExecuteCommandsInBackgroundProcess and VoiceCommand are exported from
  the timer package so main's moo() can use them
- main/timer.go becomes a thin bridge: computes room/timerUser from git
  context, then delegates to timerpkg.RunTimer / timerpkg.RunBreakTimer
- Remove executeCommandsInBackgroundProcess from mob.go (now in timer/)
